Update SimTrackSimVertexDumper to use tokens
